style : fixing phpstan errors
