+++ phpstan.neon
@@ -6,7 +6,7 @@ parameters:
 
     paths:
         -  %currentWorkingDirectory%/../src
-#        -  %currentWorkingDirectory%/../tests
+        -  %currentWorkingDirectory%/../tests
 
     bootstrap: %currentWorkingDirectory%/../vendor/autoload.php
 
@@ -21,6 +21,8 @@ parameters:
     ignoreErrors:
         - '#should end with "Interface" suffix#'
 
+        - '#Parameter \#1 \$effects of method Testomat\\TerminalColour\\AbstractStyle\:\:setEffects\(\) expects#'
+
         -
             message: '#Strict comparison using \!\=\= between null and null will always evaluate to false#'
             path: src/StyleCode.php
@@ -34,3 +36,11 @@ parameters:
         -
             message: '#Right side of \&\& is always false#'
             path: src/Util.php
+
+        # stdClass is used for test
+        - '#Parameter \#1 \$effect of method Testomat\\TerminalColour\\Contract\\Style\:\:setEffect\(\) expects array\<string\, int\|string>\|string\, stdClass given#'
+        - '#Parameter \#1 \$effect of method Testomat\\TerminalColour\\Contract\\Style\:\:unsetEffect\(\) expects array\<string\, int\|string>\|string\, stdClass given#'
+        - '#Parameter \#1 \$color of method Testomat\\TerminalColour\\Style\:\:setForeground\(\) expects int\|string\|null\, stdClass given#'
+        - '#Parameter \#1 \$color of method Testomat\\TerminalColour\\Style\:\:setBackground\(\) expects int\|string\|null\, stdClass given#'
+        - '#Parameter \#1 \$color of method Testomat\\TerminalColour\\StyleCode\:\:setForeground\(\) expects int\|string\|null\, stdClass given#'
+        - '#Parameter \#1 \$color of method Testomat\\TerminalColour\\StyleCode\:\:setBackground\(\) expects int\|string\|null\, stdClass given#'
